Add readers and writers for 3D Gaussian Splatting files
- Implemented ksplat_reader for reading .ksplat files, including compression handling and data extraction.
- Added lcc_reader and spz_reader headers and source files for future implementations.
- Created compressed_chunk writer for efficient data storage and packing.
- Developed ply_writer for writing PLY files, including header generation and binary data writing.
- Introduced DataTable structure for managing 3D data columns.
- Added necessary utility functions for data normalization and packing.
- Ensured compliance with the GNU General Public License in all new files.

+++ include/splat/lcc.h
@@ -72,7 +72,7 @@ struct ProcessUnitContext {
   std::string dataFile;
   std::string shFile;
   CompressInfo compressInfo;
-  double propertyOffset;
+  float propertyOffset;
   std::map<std::string, std::vector<float>> properties;
   std::vector<float> properties_f_rest;
 };
